Extract minion-observability (metrics and tracing), minion-taskset features.
