+++ minion/minion-features/src/main/resources/features.xml
@@ -234,19 +234,6 @@
 <!--        <bundle>mvn:org.opennms.horizon.minion.plugins/TEST-TCP-CONNECTOR/${project.version}</bundle>-->
 <!--    </feature>-->
 
-    <feature name="eclipse-stuff" version="${project.version}">
-        <bundle>mvn:org.eclipse.persistence/org.eclipse.persistence.moxy/2.5.1</bundle>
-        <bundle>mvn:org.eclipse.persistence/org.eclipse.persistence.core/2.5.1</bundle>
-        <bundle>mvn:org.eclipse.persistence/org.eclipse.persistence.asm/2.5.1</bundle>
-        <bundle>mvn:org.eclipse.persistence/org.eclipse.persistence.antlr/2.5.1</bundle>
-    </feature>
-
-    <feature name="poc-opentracing" version="${project.version}">
-        <bundle dependency="true">wrap:mvn:io.opentracing/opentracing-api/0.31.0</bundle>
-        <bundle dependency="true">wrap:mvn:io.opentracing/opentracing-noop/0.31.0</bundle>
-        <bundle dependency="true">wrap:mvn:io.opentracing/opentracing-util/0.31.0</bundle>
-    </feature>
-
 <!--    // horizon stream bundles and features needed for SNMP-->
     <feature name="poc-horizon-snmp" version="${project.version}" description="Horizon :: SNMP">
         <feature version="${project.version}">poc-grpc-horizon</feature>
