+++ minion/minion-observability/observability-tracing/src/main/resources/OSGI-INF/blueprint/tracing-noop.xml
@@ -1,0 +1,12 @@
+<blueprint xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="
+        http://www.osgi.org/xmlns/blueprint/v1.0.0
+        https://osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd
+    ">
+
+    <service interface="org.opennms.core.tracing.api.TracerWrapper">
+        <bean class="org.opennms.horizon.minion.observability.tracing.internal.NoopTracingWrapper" />
+    </service>
+
+</blueprint>
