+++ minion/minion-taskset/taskset-ipc/pom.xml
@@ -23,11 +23,6 @@
             <version>${project.version}</version>
         </dependency>
         <dependency>
-            <groupId>org.opennms.horizon.minion</groupId>
-            <artifactId>model</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
             <groupId>org.opennms.horizon.shared.ipc.grpc</groupId>
             <artifactId>contract</artifactId>
             <version>${project.version}</version>
@@ -38,8 +33,8 @@
             <version>${project.version}</version>
         </dependency>
         <dependency>
-            <groupId>org.opennms.horizon.minion</groupId>
-            <artifactId>minion-metrics</artifactId>
+            <groupId>org.opennms.horizon.minion.observability</groupId>
+            <artifactId>observability-metrics</artifactId>
             <version>${project.version}</version>
         </dependency>
     </dependencies>
