+++ minion/minion-taskset/taskset-ipc/src/main/java/org/opennms/horizon/minion/taskset/ipc/internal/LoggingWorkflowSubscriber.java
@@ -1,21 +1,21 @@
 package org.opennms.horizon.minion.taskset.ipc.internal;
 
-import org.opennms.horizon.minion.ignite.model.workflows.Workflows;
+import org.opennms.horizon.grpc.tasksets.contract.TaskSet;
 import org.opennms.horizon.minion.ipc.twin.api.TwinListener;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class LoggingWorkflowSubscriber implements TwinListener<Workflows> {
+public class LoggingWorkflowSubscriber implements TwinListener<TaskSet> {
 
   private final Logger logger = LoggerFactory.getLogger(LoggingWorkflowSubscriber.class);
 
   @Override
-  public void accept(Workflows workflowTwin) {
-    logger.info(">> Received workflow twin {}", workflowTwin);
+  public void accept(TaskSet taskSet) {
+    logger.info(">> Received task set twin {}", taskSet);
   }
 
   @Override
-  public Class<Workflows> getType() {
-    return Workflows.class;
+  public Class<TaskSet> getType() {
+    return TaskSet.class;
   }
 }
