+++ minion/minion-taskset/taskset-ipc/src/main/java/org/opennms/horizon/minion/taskset/ipc/internal/MeteredConsumer.java
@@ -7,23 +7,23 @@ import com.codahale.metrics.Metric;
 import com.codahale.metrics.MetricRegistry;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
-import org.opennms.horizon.minion.ignite.model.workflows.Results;
-import org.opennms.horizon.minion.metrics.MetricsProvider;
+import org.opennms.horizon.grpc.tasksets.contract.TaskSetResults;
+import org.opennms.horizon.minion.observability.metrics.MetricsProvider;
 import org.opennms.horizon.shared.ipc.rpc.IpcIdentity;
 
-public class MeteredConsumer implements Consumer<Results>, MetricsProvider {
+public class MeteredConsumer implements Consumer<TaskSetResults>, MetricsProvider {
 
   private final MetricRegistry metrics = new MetricRegistry();
   private final IpcIdentity identity;
-  private final Consumer<Results> delegate;
+  private final Consumer<TaskSetResults> delegate;
 
-  public MeteredConsumer(IpcIdentity identity, Consumer<Results> delegate) {
+  public MeteredConsumer(IpcIdentity identity, Consumer<TaskSetResults> delegate) {
     this.identity = identity;
     this.delegate = delegate;
   }
 
   @Override
-  public void accept(Results results) {
+  public void accept(TaskSetResults results) {
     Counter counter = get(name("minion", identity.getLocation(), identity.getId(), "result.count"), Counter::new);
     counter.inc();
     delegate.accept(results);
