+++ minion/pom.xml
@@ -23,11 +23,10 @@
         <module>snmp</module>
         <module>minion-ipc</module>
         <module>minion-grpc</module>
-        <module>minion-metrics</module>
+        <module>minion-observability</module>
         <module>minion-taskset</module>
         <module>scheduler</module>
         <module>ignite-worker</module>
-        <module>tracing-stub</module>
         <module>minion-features</module>
         <module>docker-assembly</module>
         <module>assembly</module>
