--- minion/tracing-stub/pom.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>minion</artifactId>
-        <groupId>org.opennms.horizon.minion</groupId>
-        <version>0.1.0-SNAPSHOT</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>tracing-stub</artifactId>
-    <packaging>bundle</packaging>
-    <name>OpenNms Horizon :: Minion :: Tracing Stub</name>
-
-    <description>Drop in replacement for horizon / openapi tracing</description>
-
-    <dependencies>
-        <!--            TODO MMF: Core dependency !!!-->
-        <dependency>
-            <groupId>org.opennms.horizon.core.tracing</groupId>
-            <artifactId>api</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>io.opentracing</groupId>
-            <artifactId>opentracing-noop</artifactId>
-            <version>0.31.0</version>
-        </dependency>
-    </dependencies>
-
-</project>
