--- minion/tracing-stub/src/main/resources/OSGI-INF/blueprint/tracing-noop.xml
@@ -1,12 +1,0 @@
-<blueprint xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="
-        http://www.osgi.org/xmlns/blueprint/v1.0.0
-        https://osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd
-    ">
-
-    <service interface="org.opennms.core.tracing.api.TracerWrapper">
-        <bean class="org.opennms.poc.tracing.NoopTracingWrapper" />
-    </service>
-
-</blueprint>
